minor fixes to 6.steps

+++ 6.steps/clean_folder.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
-rm -rf *.log *.map *.ped *.log *.vcf *.raw *.nosex *.kinship *.irem *.pdf *.results *.png *.vcf.gz
-rm -r data/
+rm -rf *.log *.map *.ped *.log *.vcf *.raw *.nosex *.kinship *.irem *.pdf *.results *.png *.vcf.gz *.results.gz
+
+if [ -d "data" ] 
+then
+	rm -r data/
+fi
